Revert "refactor: add store to provider"
This reverts commit 88bf5c4a79e65418e5f1c444e12a9c739d4e8b3c.
