+++ apps/nextjs/src/components/AppComponents/Chat/chat-lessonPlanDisplay.stories.tsx
@@ -3,7 +3,6 @@ import type { Meta, StoryObj } from "@storybook/react";
 import type { ChatContextProps } from "@/components/ContextProviders/ChatProvider";
 import { chromaticParams } from "@/storybook/chromatic";
 import { ChatDecorator } from "@/storybook/decorators/ChatDecorator";
-import { ModerationStoreDecorator } from "@/storybook/decorators/ModerationStoreDecorator";
 
 import LessonPlanDisplay from "./chat-lessonPlanDisplay";
 
@@ -27,7 +26,7 @@ const meta = {
   title: "Components/LessonPlan/LessonPlanDisplay",
   component: LessonPlanDisplay,
   tags: ["autodocs"],
-  decorators: [ChatDecorator, ModerationStoreDecorator],
+  decorators: [ChatDecorator],
   args: {
     documentContainerRef: { current: null },
     chatEndRef: undefined,
@@ -65,8 +64,6 @@ export const WithModeration: Story = {
   parameters: {
     chatContext: {
       ...chatContext,
-    },
-    moderationStoreContext: {
       lastModeration: {
         id: "123",
         categories: ["l/strong-language"],
